ownership_employee/conf/urls.py
from django.urls import path, include
from django.contrib import admin
from . import settings

urlpatterns = [
    path('admin/', admin.site.urls),
    path('members/', include('members.urls')),
    path('', include('dashboard.urls')),
    path('dashboard/', include('dashboard.urls')),
]

if settings.API:
    from rest_framework_simplejwt.views import (
        TokenObtainPairView,
        TokenRefreshView,
    )

    urlpatterns = [
        path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
        path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
        path('api/', include('members.api.urls')),
        path('api', include('rest_framework.urls', namespace='rest_framework')),
    ] + urlpatterns


if settings.DEBUG:
    import debug_toolbar
    from django.conf.urls.static import static
    urlpatterns = [
        path(r'__debug__/', include(debug_toolbar.urls)),
    ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT) + urlpatterns
